trips_be/requirements.txt
asgiref==3.2.10
boto3==1.16.5
botocore==1.19.5
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
jmespath==0.10.0
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
